pyproject.toml
[tool.poetry]
name = "forms-app"
version = "0.1.0"
description = ""
authors = ["Анна Полякова <Sova040888@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.0"
pymongo = "^4.6.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.1"
spectree = "^1.2.8"
phonenumbers = "^8.13.25"
validate-email-address = "^1"
pydantic-settings = "^2.1.0"
pytest = "^7.4.3"
pytest-lazy-fixture = "^0.6.3"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
